# BackEnd/docs/swagger.yaml
basePath: /
definitions:
...
  web_doscom_internal_database_model.GalleryResponse:
    properties:
      description:
        type: string
      event_date:
        type: string
      file_upload_id:
        type: integer
      gallery_name:
        type: string
      gallery_type:
        type: string
      id:
        type: integer
      id_users:
        type: integer
    type: object
...
info:
  contact: {}
  description: API Documentation Web Doscom
  title: Web Doscom API
  version: "1.0"
paths:
...
  /api/v1/gallery/:
    get:
      consumes:
      - application/json
      description: Mengambil data gallery berdasarkan tipe tertentu
      parameters:
      - description: Gallery type (fun, proker, achievment, work, activity, blog,
          pengurus, etc)
        in: query
        name: type
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetch gallery data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized, role tidak valid
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to fetch gallery data
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get Gallery By Type
      tags:
      - Gallery
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple images or videos to MinIO storage with metadata.
        All files stored in MinIO bucket 'doscom-uploads/gallery'.
      parameters:
      - description: Type of gallery (fun, proker, achievment, work, activity, blog,
          pengurus, etc)
        in: formData
        name: gallery_type
        required: true
        type: string
      - description: Description of gallery
        in: formData
        name: description
        required: true
        type: string
      - description: Event date (YYYY-MM-DD)
        in: formData
        name: event_date
        required: true
        type: string
      - collectionFormat: csv
        description: Upload multiple image/video files
        in: formData
        items:
          type: file
        name: files
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Successfully uploaded to MinIO
          schema:
            $ref: '#/definitions/web_doscom_internal_database_model.GalleryResponse'
        "400":
          description: Bad request - missing fields or invalid files
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error - upload failed
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload gallery images/videos
      tags:
      - Gallery
  /api/v1/gallery/{id}:
    delete:
      consumes:
      - application/json
      description: Delete gallery data dan file nya berdasarkan ID
      parameters:
      - description: Gallery ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete Gallery
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Gallery not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete gallery by ID
      tags:
      - Gallery
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"